backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# WebSocket
python-socketio==5.10.0

# PLC Communication
python-snap7==1.3

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Reports
reportlab==4.0.8
openpyxl==3.1.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
